[WEJBHTTP-141] Refactoring - Moving RequestBuilder.Invocation type to RequestType.
